Mihentai (ID): update language and icon
