+++ src/id/mihentai/build.gradle
@@ -2,8 +2,8 @@ ext {
     extName = 'Mihentai'
     extClass = '.Mihentai'
     themePkg = 'mangathemesia'
-    baseUrl = 'https://mihentai.com'
-    overrideVersionCode = 2
+    baseUrl = 'https://mihentai.net'
+    overrideVersionCode = 3
     isNsfw = true
 }
